287 Move sidebar selectable item into its own module
